agents/requirements_langchain.txt
boto3>=1.34.0
botocore>=1.34.0
numpy>=1.24.0
scikit-learn>=1.3.0
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0
pyyaml>=6.0.0